requirements.txt
Flask==1.1.2
google-api-core==1.22.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.4.1
google-cloud-speech==1.3.2
google-cloud-storage==1.31.0
google-cloud-texttospeech==1.0.1
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.51.0
h5py==2.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
librosa==0.8.0
matplotlib==3.3.1
numba==0.51.2
numpy==1.18.5
pandas==1.1.1
pandocfilters==1.4.2
Pillow==7.2.0
QtPy==1.9.0
scikit-learn==0.23.2
scipy==1.4.1
SoundFile==0.10.3.post1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Theano==1.0.5
urllib3==1.25.10
gunicorn==19.0.0
pydub==0.24.1
scikit-learn==0.23.2
audioread==2.1.8